refactor: aws clients, method params, file naming
