+++ .env.example
@@ -1,9 +1,16 @@
+# DB
 DB_URL=
 POSTGRES_USER=
 POSTGRES_PASSWORD=
 POSTGRES_HOST=
 POSTGRES_PORT=
 POSTGRES_DB=
+
 REFRESH_TOKEN_SECRET=
 ACCESS_TOKEN_SECRET=
+
 SENTRY_DSN=
+
+# AWS
+AWS_ACCESS_KEY_ID=
+AWS_SECRET_ACCESS_KEY=
