+++ src/controllers/transcription.py
@@ -3,6 +3,7 @@ import uuid
 import http
 import requests
 
+from os import environ as env
 from fastapi import APIRouter, Depends
 from botocore.exceptions import ClientError
 
@@ -25,18 +26,20 @@ transcription_router = APIRouter(
   response_model=GenericResponseModel
 )
 async def transcribe_audio(s3_object_name: str, language_code = "id-ID"):
-  # s3_client = AWSS3Client().get_client()
   service = TranscriptionService()
   transcribe_client = AWSTranscribeClient().get_client()
 
   filename, file_format = s3_object_name.split(".")
   generated_job_name = service.generate_job_name()
 
+  BUCKET_NAME = env.get("AWS_BUCKET_NAME")
+  file_uri = service.generate_file_uri(bucket_name=BUCKET_NAME, filename=filename, extension=file_format)
+
   try:
     response = service.transcribe_file(
       transcribe_client=transcribe_client,
       job_name=generated_job_name,
-      filename=filename,
+      file_uri=file_uri,
       file_format=file_format,
       language_code=language_code
     ) 
@@ -46,7 +49,7 @@ async def transcribe_audio(s3_object_name: str, language_code = "id-ID"):
       message="Successfully created audio transcription",
       data=response
     )
-  except ClientError as err:
+  except ClientError:
     return GenericResponseModel(
       status_code=http.HTTPStatus.REQUEST_TIMEOUT,
       error="Timeout while processing transcription job.",
